Changes in the project structure
Plus minor improvements
