+++ gtd.py
@@ -1,8 +1,25 @@
-import interface
-from utilities import *
+# -*- coding: utf-8 -*- 
+"""Copyright (C) 2016  Łukasz Bednarski
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program. If not, see http://www.gnu.org/licenses/.
+"""
+import src.interface as interface
+
+__version__ = 0.1
 
 if __name__ == '__main__':
-    db = load('tasks.xml')
+    db = interface.io.load('tasks.xml')
     ui = interface.Interface(db)
     ui.start()
-    save(db)
+    interface.io.save(db)
